requirements.txt
locust
python-dotenv
pytest
requests